Consolidate SlocHtmlTest into sloc_test.py
